requirements.txt
pyodbc
python-dotenv
PyYAML
openai>=1.30.0
